+++ CHANGES.rst
@@ -2,6 +2,7 @@ Unreleased
 ----------
 
 - Bugfix: Handle custom scalar that subclass :class:`py_gql.schema.ScalarType` during execution.
+- Bugfix: Fix inheritance for :class:`py_gql.schema.RegexType`.
 
 0.1.0
 -----